Implement functionality for adding a counter
Simply increment numViews and display a new view with this index.
When viewDidLoad is called, it will attempt to load a Counter from
file counterN on disk, where N is the new index.  When it doesn't find the file,
it will call reset, which will initialize a new Counter with default values and
save it to disk.

Issue #5