Fix indexing bug
numStoredCounters should be incremented when a new counter is saved to disk, which can occur at times
other than when "+" is pushed (e.g., if the last counter is deleted).

In addition, C's % operator returns a negative number if the divisor is negative, so we need to check for this manually.